Dataframe for the expectec correlation

Dataframe for the expectec correlation

data_frame_expected_cov(
  sigma_Z,
  the_radius,
  the_directions,
  the_scale = 1,
  direction = TRUE
)

Arguments

sigma_Z
the standard error of the theoritical distribution

the_radius
the radius of the window of the moving average

the_directions
the directions we consider in a list ex : list(c(0,1), c(1, 1))

the_scale
how many km per pixels ?

direction
TRUE by default. reduce the direction into its director vector

Value

A dataframe with seven colomns : the distance between two points in pixels unit, this distance in km unit, the value of the expected covariance and expected correlation the considered direction, and the radius of the window of the moving average (in pixels and then km unit).

Examples

data_frame_expected_cov(1, c(1), list(c(0,1), c(1,1), c(1, 2)))

#>    Distance_pixels Distance_km Expected_covariance Expected_correlation
#> 1         0.000000    0.000000          0.11111111            1.0000000
#> 2         1.000000    1.000000          0.07407407            0.6666667
#> 3         2.000000    2.000000          0.03703704            0.3333333
#> 4         3.000000    3.000000          0.00000000            0.0000000
#> 5         4.000000    4.000000          0.00000000            0.0000000
#> 6         5.000000    5.000000          0.00000000            0.0000000
#> 7         6.000000    6.000000          0.00000000            0.0000000
#> 8         7.000000    7.000000          0.00000000            0.0000000
#> 9         8.000000    8.000000          0.00000000            0.0000000
#> 10        9.000000    9.000000          0.00000000            0.0000000
#> 11        0.000000    0.000000          0.11111111            1.0000000
#> 12        1.414214    1.414214          0.04938272            0.4444444
#> 13        2.828427    2.828427          0.01234568            0.1111111
#> 14        4.242641    4.242641          0.00000000            0.0000000
#> 15        5.656854    5.656854          0.00000000            0.0000000
#> 16        7.071068    7.071068          0.00000000            0.0000000
#> 17        8.485281    8.485281          0.00000000            0.0000000
#> 18        9.899495    9.899495          0.00000000            0.0000000
#> 19       11.313708   11.313708          0.00000000            0.0000000
#> 20       12.727922   12.727922          0.00000000            0.0000000
#> 21        0.000000    0.000000          0.11111111            1.0000000
#> 22        2.236068    2.236068          0.02469136            0.2222222
#> 23        4.472136    4.472136          0.00000000            0.0000000
#> 24        6.708204    6.708204          0.00000000            0.0000000
#> 25        8.944272    8.944272          0.00000000            0.0000000
#> 26       11.180340   11.180340          0.00000000            0.0000000
#> 27       13.416408   13.416408          0.00000000            0.0000000
#> 28       15.652476   15.652476          0.00000000            0.0000000
#>    Direction Radius_pixels Radius_km
#> 1      0 - 1             1         1
#> 2      0 - 1             1         1
#> 3      0 - 1             1         1
#> 4      0 - 1             1         1
#> 5      0 - 1             1         1
#> 6      0 - 1             1         1
#> 7      0 - 1             1         1
#> 8      0 - 1             1         1
#> 9      0 - 1             1         1
#> 10     0 - 1             1         1
#> 11     1 - 1             1         1
#> 12     1 - 1             1         1
#> 13     1 - 1             1         1
#> 14     1 - 1             1         1
#> 15     1 - 1             1         1
#> 16     1 - 1             1         1
#> 17     1 - 1             1         1
#> 18     1 - 1             1         1
#> 19     1 - 1             1         1
#> 20     1 - 1             1         1
#> 21     1 - 2             1         1
#> 22     1 - 2             1         1
#> 23     1 - 2             1         1
#> 24     1 - 2             1         1
#> 25     1 - 2             1         1
#> 26     1 - 2             1         1
#> 27     1 - 2             1         1
#> 28     1 - 2             1         1